Move templates section and logagentspec section
