+++ pkg/api/v1beta3/dynakube/dynakube_types.go
@@ -164,6 +164,16 @@ type DynaKubeSpec struct { //nolint:revive
 	LogMonitoring LogMonitoringSpec `json:"logMonitoring,omitempty"`
 }
 
+type TemplatesSpec struct {
+
+	// +kubebuilder:validation:Optional
+	OpenTelemetryCollector OpenTelemetryCollectorSpec `json:"openTelemetryCollector,omitempty"`
+	// +kubebuilder:validation:Optional
+	ExtensionExecutionController ExtensionExecutionControllerSpec `json:"extensionExecutionController,omitempty"`
+	// +kubebuilder:validation:Optional
+	LogAgent LogAgentSpec `json:"logAgent,omitempty"`
+}
+
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 
 // DynaKubeList contains a list of DynaKube
